Split retry code to retry.go
